pyproject.toml
[project]
name = "autotag"
version = "0.1.0"
description = "Hybrid rules + ML + LLM stub auto-tagging service"
authors = [{name = "Sindibad ProdEng"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.24",
    "SQLAlchemy>=2.0",
    "pydantic>=2.3",
    "pydantic-settings>=2.0",
    "scikit-learn>=1.3",
    "joblib>=1.3",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "httpx>=0.25",
    "black>=23.9",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://example.com"

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["autotag"]
addopts = "-q"
